+++ core/objects/src/main/kotlin/app/aaps/core/objects/extensions/TemporaryBasalExtension.kt
@@ -102,7 +102,7 @@ fun TB.iobCalc(
     profile: Profile,
     lastAutosensResult: AutosensResult,
     exerciseMode: Boolean,
-    halfBasalExerciseTarget: Int,
+    halfBasalExerciseTarget: Double,
     isTempTarget: Boolean,
     insulinInterface: Insulin
 ): IobTotal {
